ima: Add support for measurement list snapshots
Introduce the ability of taking snapshots of the IMA measurement list.
Making a snapshot means moving the current content of the measurement list
to a separate location, and allowing users to read and delete it. This
causes the measurement list to be atomically truncated before new
measurements can be added. Only one snapshot can be created at a time.

User space is responsible to concatenate the snapshots following the
temporal order in which they were taken, together with the current
measurement list. Then, it can send the collected data to the remote
verifiers.

The benefit of this solution is the ability to free precious kernel memory,
in exchange of delegating user space to reconstruct the full measurement
list from the chunks. No trust needs to be given to user space, since the
integrity of the measurement list is protected by the TPM.

By default, taking a snapshot of the measurement list does not alter the
hash table. When a snapshot is created, IMA is still able to detect
collisions on the entries moved to the snapshots and eventually deleted
(the measurement entries are partially kept in kernel memory for that
reason, only the template data is freed).

Introduce the new kernel option ima_flush_htable to remove the measurement
entries from the hash table at the time a snapshot is taken, and to free
their memory when a snapshot is deleted.

Then, introduce ascii_runtime_measurements_snap_<algo> and
binary_runtime_measurement_snap_<algo> interfaces to manage the snapshot
lifecycle. Use 'echo 0 > <IMA interface>' and 'echo 1 > <IMA interface>' to
respectively create and delete a snapshot. The taken snapshot can be read
like the other measurement interfaces.

The ima_measure_sem semaphore has been introduced to protect access to the
measurement list and the snapshot. The open method of all the measurement
interfaces has been extended to try to lock the new semaphore, for read or
for write depending on the operation to perform (write: create/delete the
snapshot, read: read the current measurement list or the snapshot). In
addition, the open method checks if the current process has the
CAP_SYS_ADMIN capability, when a write is requested.

Multiple readers are allowed to access the measurement list interfaces
concurrently (same behavior as before this patch), while only one writer is
allowed to open either the regular measurement list interfaces or the new
ones for snapshot. In this way, readers don't see the measurement list
suddenly truncated and don't access the snapshot while it is being deleted.

Note: This code derives from the Alt-IMA Huawei project, and is being
      released under the dual license model (GPL-2.0 OR MIT).

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>